refactor: move sudt to separated module
